+++ src/app/components/content-harvesters-spaces.js
@@ -23,6 +23,7 @@ import {
   initDestroyableCache,
 } from 'onedata-gui-common/utils/destroyable-computed';
 import { getNameWithConflictLabel } from 'onedata-gui-common/components/name-conflict';
+import { reads } from '@ember/object/computed';
 
 export default Component.extend(I18n, GlobalActions, {
   classNames: ['content-harvesters-spaces'],
@@ -58,14 +59,17 @@ export default Component.extend(I18n, GlobalActions, {
       } else {
         return resolve([]);
       }
-    })),
+    }
+  )),
+
+  harvesterSpaces: reads('harvesterSpacesProxy.content'),
 
   /**
    * @type {ComputedProperty<Array<SpaceListItem>>}
    */
-  spaceItems: computed('harvesterSpacesProxy.[]', function spaceItems() {
+  spaceItems: computed('harvesterSpaces.[]', function spaceItems() {
     const harvester = this.get('harvester');
-    const spaces = this.get('harvesterSpacesProxy.content') || [];
+    const spaces = this.harvesterSpaces || [];
     return spaces.map(space => SpaceListItem.create({
       ownerSource: this,
       parentHarvester: harvester,
@@ -77,15 +81,15 @@ export default Component.extend(I18n, GlobalActions, {
    * @type {ComputedProperty<Array<SpaceListItem>>}
    */
   filteredSpaceItems: computed(
-    'spaceItems.[]',
+    'spaceItems.@each.{name,conflictLabel}',
     'searchString',
     function filteredSpaceItems() {
       return this.spaceItems.filter(item => {
         const searchableName = getNameWithConflictLabel(
-          item.record.name,
-          item.record.conflictLabel
+          item.name,
+          item.conflictLabel
         );
-        return searchableName.toLowerCase().includes(this.searchString.toLowerCase());
+        return searchableName?.toLowerCase().includes(this.searchString.toLowerCase());
       });
     }
   ),
@@ -166,6 +170,16 @@ const SpaceListItem = ResourceListItem.extend(OwnerInjector, {
    * @virtual
    */
   parentHarvester: undefined,
+
+  /**
+   * @type {ComputedProperty<string>}
+   */
+  name: reads('record.name'),
+
+  /**
+   * @type {ComputedProperty<string>}
+   */
+  conflictLabel: reads('record.conflictLabel'),
 
   /**
    * @override